+++ tests/testthat/test-assertions.R
@@ -523,8 +523,8 @@ test_that("assert_vars Test 31: error if some elements of `arg` are not unquoted
 })
 
 # assert_order_vars ----
-## Test 32: error if `arg` is not a list variable names or `desc()` ----
-test_that("assert_order_vars Test 32: error if `arg` is not a list variable names or `desc()`", {
+## Test 32: error if `arg` is not a list ----
+test_that("assert_order_vars Test 32: error if `arg` is not a list", {
   expect_error(
     assert_order_vars(arg <- TRUE),
     regexp =
@@ -545,23 +545,32 @@ test_that("assert_order_vars Test 32: error if `arg` is not a list variable name
   )
 })
 
-## Test 33: no error if valid argument ----
-test_that("assert_order_vars Test 33: no error if valid argument", {
+## Test 33: error if invalid elements ----
+test_that("assert_order_vars Test 33: error if invalid elements", {
+  expect_error(
+    assert_order_vars(arg <- exprs(USUBJID, !!mean)),
+    regexp = "Each element of `arg` must be an expression but the following are not:",
+    fixed = TRUE
+  )
+})
+
+## Test 34: no error if valid argument ----
+test_that("assert_order_vars Test 34: no error if valid argument", {
   expect_invisible(
     assert_order_vars(arg <- exprs(USUBJID, desc(PARAMCD), -abs(AVAL)))
   )
 })
 
-## Test 34: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_order_vars Test 34: no error if `arg` is NULL and optional is TRUE", {
+## Test 35: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_order_vars Test 35: no error if `arg` is NULL and optional is TRUE", {
   expect_invisible(
     assert_order_vars(arg <- NULL, optional = TRUE)
   )
 })
 
 # assert_integer_scalar ----
-## Test 35: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_integer_scalar Test 35: no error if `arg` is NULL and optional is TRUE", {
+## Test 36: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_integer_scalar Test 36: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_integer_scalar(arg, optional = TRUE)
   }
@@ -571,8 +580,8 @@ test_that("assert_integer_scalar Test 35: no error if `arg` is NULL and optional
   )
 })
 
-## Test 36: error if chosen subset not in subsets ----
-test_that("assert_integer_scalar Test 36: error if chosen subset not in subsets", {
+## Test 37: error if chosen subset not in subsets ----
+test_that("assert_integer_scalar Test 37: error if chosen subset not in subsets", {
   example_fun <- function(arg) {
     assert_integer_scalar(arg, subset = "infinity")
   }
@@ -582,8 +591,8 @@ test_that("assert_integer_scalar Test 36: error if chosen subset not in subsets"
   )
 })
 
-## Test 37: no error if `arg` is in selected subset ----
-test_that("assert_integer_scalar Test 37: no error if `arg` is in selected subset", {
+## Test 38: no error if `arg` is in selected subset ----
+test_that("assert_integer_scalar Test 38: no error if `arg` is in selected subset", {
   example_fun <- function(arg) {
     assert_integer_scalar(arg, subset = "positive")
   }
@@ -593,8 +602,8 @@ test_that("assert_integer_scalar Test 37: no error if `arg` is in selected subse
   )
 })
 
-## Test 38: error if `arg` is not an integer scalar ----
-test_that("assert_integer_scalar Test 38: error if `arg` is not an integer scalar", {
+## Test 39: error if `arg` is not an integer scalar ----
+test_that("assert_integer_scalar Test 39: error if `arg` is not an integer scalar", {
   example_fun <- function(arg) {
     assert_integer_scalar(arg)
   }
@@ -608,8 +617,8 @@ test_that("assert_integer_scalar Test 38: error if `arg` is not an integer scala
 })
 
 # assert_numeric_vector ----
-## Test 39: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_numeric_vector Test 39: no error if `arg` is NULL and optional is TRUE", {
+## Test 40: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_numeric_vector Test 40: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_numeric_vector(arg, optional = TRUE)
   }
@@ -620,8 +629,8 @@ test_that("assert_numeric_vector Test 39: no error if `arg` is NULL and optional
 })
 
 # assert_integer_scalar ----
-## Test 40: error if `arg` is not an integer scalar ----
-test_that("assert_integer_scalar Test 40: error if `arg` is not an integer scalar", {
+## Test 41: error if `arg` is not an integer scalar ----
+test_that("assert_integer_scalar Test 41: error if `arg` is not an integer scalar", {
   example_fun <- function(arg) {
     assert_numeric_vector(arg)
   }
@@ -634,8 +643,8 @@ test_that("assert_integer_scalar Test 40: error if `arg` is not an integer scala
 })
 
 # assert_s3_class ----
-## Test 41: error if `arg` is not an object of a specific class S3 ----
-test_that("assert_s3_class Test 41: error if `arg` is not an object of a specific class S3", {
+## Test 42: error if `arg` is not an object of a specific class S3 ----
+test_that("assert_s3_class Test 42: error if `arg` is not an object of a specific class S3", {
   example_fun <- function(arg) {
     assert_s3_class(arg, "factor")
   }
@@ -643,8 +652,8 @@ test_that("assert_s3_class Test 41: error if `arg` is not an object of a specifi
   expect_error(example_fun("test"))
 })
 
-## Test 42: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_s3_class Test 42: no error if `arg` is NULL and optional is TRUE", {
+## Test 43: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_s3_class Test 43: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_s3_class(arg, class = "factor", optional = TRUE)
   }
@@ -654,8 +663,8 @@ test_that("assert_s3_class Test 42: no error if `arg` is NULL and optional is TR
   )
 })
 
-## Test 43: error if `arg` is NULL and optional is FALSE ----
-test_that("assert_s3_class Test 43: error if `arg` is NULL and optional is FALSE", {
+## Test 44: error if `arg` is NULL and optional is FALSE ----
+test_that("assert_s3_class Test 44: error if `arg` is NULL and optional is FALSE", {
   example_fun <- function(arg) {
     assert_s3_class(arg, class = "factor", optional = FALSE)
   }
@@ -667,8 +676,8 @@ test_that("assert_s3_class Test 43: error if `arg` is NULL and optional is FALSE
   )
 })
 
-## Test 44: no error if `arg` is an object of a specific class S3 ----
-test_that("assert_s3_class Test 44: no error if `arg` is an object of a specific class S3", {
+## Test 45: no error if `arg` is an object of a specific class S3 ----
+test_that("assert_s3_class Test 45: no error if `arg` is an object of a specific class S3", {
   example_fun <- function(arg) {
     assert_s3_class(arg, "factor")
   }
@@ -677,8 +686,8 @@ test_that("assert_s3_class Test 44: no error if `arg` is an object of a specific
 })
 
 # assert_list_of ----
-## Test 45: error if `arg` is not a list of specific class S3 objects ----
-test_that("assert_list_of Test 45: error if `arg` is not a list of specific class S3 objects", {
+## Test 46: error if `arg` is not a list of specific class S3 objects ----
+test_that("assert_list_of Test 46: error if `arg` is not a list of specific class S3 objects", {
   example_fun <- function(arg) {
     assert_list_of(arg, "factor")
   }
@@ -686,8 +695,8 @@ test_that("assert_list_of Test 45: error if `arg` is not a list of specific clas
   expect_error(example_fun(list("test")))
 })
 
-## Test 46: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_list_of Test 46: no error if `arg` is NULL and optional is TRUE", {
+## Test 47: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_list_of Test 47: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_list_of(arg, class = "factor", optional = TRUE)
   }
@@ -697,8 +706,8 @@ test_that("assert_list_of Test 46: no error if `arg` is NULL and optional is TRU
   )
 })
 
-## Test 47: error if `arg` is NULL and optional is FALSE ----
-test_that("assert_list_of Test 47: error if `arg` is NULL and optional is FALSE", {
+## Test 48: error if `arg` is NULL and optional is FALSE ----
+test_that("assert_list_of Test 48: error if `arg` is NULL and optional is FALSE", {
   example_fun <- function(arg) {
     assert_list_of(arg, class = "factor", optional = FALSE)
   }
@@ -710,8 +719,8 @@ test_that("assert_list_of Test 47: error if `arg` is NULL and optional is FALSE"
   )
 })
 
-## Test 48: no error if `arg` is a list of specific class S3 objects ----
-test_that("assert_list_of Test 48: no error if `arg` is a list of specific class S3 objects", {
+## Test 49: no error if `arg` is a list of specific class S3 objects ----
+test_that("assert_list_of Test 49: no error if `arg` is a list of specific class S3 objects", {
   example_fun <- function(arg) {
     assert_list_of(arg, "factor")
   }
@@ -723,8 +732,8 @@ test_that("assert_list_of Test 48: no error if `arg` is a list of specific class
   )
 })
 
-## Test 49: error if `arg` is not a named list (no elements named) ----
-test_that("assert_list_of Test 49: error if `arg` is not a named list (no elements named)", {
+## Test 50: error if `arg` is not a named list (no elements named) ----
+test_that("assert_list_of Test 50: error if `arg` is not a named list (no elements named)", {
   expect_error(
     assert_list_of(mylist <- list(1, 2, 3), class = "numeric", named = TRUE),
     paste(
@@ -735,8 +744,8 @@ test_that("assert_list_of Test 49: error if `arg` is not a named list (no elemen
   )
 })
 
-## Test 50: error if `arg` is not a named list (some elements named) ----
-test_that("assert_list_of Test 50: error if `arg` is not a named list (some elements named)", {
+## Test 51: error if `arg` is not a named list (some elements named) ----
+test_that("assert_list_of Test 51: error if `arg` is not a named list (some elements named)", {
   expect_error(
     assert_list_of(mylist <- list(1, 2, 3, d = 4), class = "numeric", named = TRUE),
     paste(
@@ -747,16 +756,35 @@ test_that("assert_list_of Test 50: error if `arg` is not a named list (some elem
   )
 })
 
-## Test 51: no error if `arg` is a named list ----
-test_that("assert_list_of Test 51: no error if `arg` is a named list", {
+## Test 52: no error if `arg` is a named list ----
+test_that("assert_list_of Test 52: no error if `arg` is a named list", {
   expect_invisible(
     assert_list_of(mylist <- list(a = 1, b = 2, c = 3), class = "numeric", named = TRUE)
   )
 })
 
+# assert_named ----
+## Test 53: no error if arg is NULL and optional = TRUE ----
+test_that("assert_named Test 53: no error if arg is NULL and optional = TRUE", {
+  expect_invisible(assert_named(arg <- NULL, optional = TRUE))
+})
+
+## Test 54: error if no elements are named ----
+test_that("assert_named Test 54: error if no elements are named", {
+  expect_error(
+    assert_named(arg <- c(1, 2)),
+    regexp = paste(
+      "All elements of `arg` must be named.",
+      "No element is named.",
+      sep = "\n"
+    ),
+    fixed = TRUE
+  )
+})
+
 # assert_named_exprs ----
-## Test 52: error if `arg` is not a named list of expressions ----
-test_that("assert_named_exprs Test 52: error if `arg` is not a named list of expressions", {
+## Test 55: error if `arg` is not a named list of expressions ----
+test_that("assert_named_exprs Test 55: error if `arg` is not a named list of expressions", {
   example_fun <- function(arg) {
     assert_named_exprs(arg)
   }
@@ -770,8 +798,8 @@ test_that("assert_named_exprs Test 52: error if `arg` is not a named list of exp
   expect_error(example_fun(arg))
 })
 
-## Test 53: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_named_exprs Test 53: no error if `arg` is NULL and optional is TRUE", {
+## Test 56: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_named_exprs Test 56: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_named_exprs(arg, optional = TRUE)
   }
@@ -781,8 +809,8 @@ test_that("assert_named_exprs Test 53: no error if `arg` is NULL and optional is
   )
 })
 
-## Test 54: no error if `arg` is a named list of expressions ----
-test_that("assert_named_exprs Test 54: no error if `arg` is a named list of expressions", {
+## Test 57: no error if `arg` is a named list of expressions ----
+test_that("assert_named_exprs Test 57: no error if `arg` is a named list of expressions", {
   example_fun <- function(arg) {
     assert_named_exprs(arg)
   }
@@ -795,8 +823,8 @@ test_that("assert_named_exprs Test 54: no error if `arg` is a named list of expr
 })
 
 # assert_function ----
-## Test 55: error if `arg` is not a function ----
-test_that("assert_function Test 55: error if `arg` is not a function", {
+## Test 58: error if `arg` is not a function ----
+test_that("assert_function Test 58: error if `arg` is not a function", {
   example_fun <- function(arg) {
     assert_function(arg)
   }
@@ -805,8 +833,8 @@ test_that("assert_function Test 55: error if `arg` is not a function", {
   expect_error(example_fun())
 })
 
-## Test 56: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_function Test 56: no error if `arg` is NULL and optional is TRUE", {
+## Test 59: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_function Test 59: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_function(arg, optional = TRUE)
   }
@@ -816,8 +844,8 @@ test_that("assert_function Test 56: no error if `arg` is NULL and optional is TR
   )
 })
 
-## Test 57: no error if `arg` is a function with all parameters defined ----
-test_that("assert_function Test 57: no error if `arg` is a function with all parameters defined", {
+## Test 60: no error if `arg` is a function with all parameters defined ----
+test_that("assert_function Test 60: no error if `arg` is a function with all parameters defined", {
   example_fun <- function(arg) {
     assert_function(arg, params = c("x"))
   }
@@ -825,8 +853,8 @@ test_that("assert_function Test 57: no error if `arg` is a function with all par
   expect_invisible(example_fun(mean))
 })
 
-## Test 58: error if  `params`  is missing with no default ----
-test_that("assert_function Test 58: error if  `params`  is missing with no default", {
+## Test 61: error if  `params`  is missing with no default ----
+test_that("assert_function Test 61: error if  `params`  is missing with no default", {
   example_fun <- function(arg) {
     assert_function(arg, params = c("x"))
   }
@@ -842,8 +870,8 @@ test_that("assert_function Test 58: error if  `params`  is missing with no defau
 
 
 # assert_function_param ----
-## Test 59: no error if `arg` is a parameter of a function ----
-test_that("assert_function_param Test 59: no error if `arg` is a parameter of a function", {
+## Test 62: no error if `arg` is a parameter of a function ----
+test_that("assert_function_param Test 62: no error if `arg` is a parameter of a function", {
   hello <- function(name) {
     print(sprintf("Hello %s", name))
   }
@@ -851,8 +879,8 @@ test_that("assert_function_param Test 59: no error if `arg` is a parameter of a 
   expect_invisible(assert_function_param("hello", "name"))
 })
 
-## Test 60: error if expected function parameters are missing ----
-test_that("assert_function_param Test 60: error if expected function parameters are missing", {
+## Test 63: error if expected function parameters are missing ----
+test_that("assert_function_param Test 63: error if expected function parameters are missing", {
   hello <- function(name) {
     print(sprintf("Hello %s", name))
   }
@@ -862,8 +890,8 @@ test_that("assert_function_param Test 60: error if expected function parameters 
 })
 
 # assert_unit ----
-## Test 61: no error if the parameter is provided in the expected unit ----
-test_that("assert_unit Test 61: no error if the parameter is provided in the expected unit", {
+## Test 64: no error if the parameter is provided in the expected unit ----
+test_that("assert_unit Test 64: no error if the parameter is provided in the expected unit", {
   advs <- tibble::tribble(
     ~USUBJID, ~VSTESTCD, ~VSTRESN, ~VSSTRESU, ~PARAMCD, ~AVAL,
     "P01",    "WEIGHT",      80.1, "kg",      "WEIGHT",  80.1,
@@ -875,8 +903,8 @@ test_that("assert_unit Test 61: no error if the parameter is provided in the exp
   )
 })
 
-## Test 62: error if there are multiple units in the input dataset ----
-test_that("assert_unit Test 62: error if there are multiple units in the input dataset", {
+## Test 65: error if there are multiple units in the input dataset ----
+test_that("assert_unit Test 65: error if there are multiple units in the input dataset", {
   advs <- tibble::tribble(
     ~USUBJID, ~VSTESTCD, ~VSTRESN, ~VSSTRESU, ~PARAMCD, ~AVAL,
     "P01",    "WEIGHT",      80.1, "kg",      "WEIGHT",  80.1,
@@ -888,8 +916,8 @@ test_that("assert_unit Test 62: error if there are multiple units in the input d
   )
 })
 
-## Test 63: error if unexpected unit in the input dataset ----
-test_that("assert_unit Test 63: error if unexpected unit in the input dataset", {
+## Test 66: error if unexpected unit in the input dataset ----
+test_that("assert_unit Test 66: error if unexpected unit in the input dataset", {
   advs <- tibble::tribble(
     ~USUBJID, ~VSTESTCD, ~VSTRESN, ~VSSTRESU, ~PARAMCD, ~AVAL,
     "P01",    "WEIGHT",      80.1, "kg",      "WEIGHT",  80.1,
@@ -902,8 +930,8 @@ test_that("assert_unit Test 63: error if unexpected unit in the input dataset", 
 })
 
 # assert_param_does_not_exist ----
-## Test 64: error if parameter exists in the input dataset ----
-test_that("assert_param_does_not_exist Test 64: error if parameter exists in the input dataset", {
+## Test 67: error if parameter exists in the input dataset ----
+test_that("assert_param_does_not_exist Test 67: error if parameter exists in the input dataset", {
   advs <- tibble::tribble(
     ~USUBJID, ~VSTESTCD, ~VSTRESN, ~VSSTRESU, ~PARAMCD, ~AVAL,
     "P01",    "WEIGHT",      80.1, "kg",      "WEIGHT",  80.1,
@@ -915,8 +943,8 @@ test_that("assert_param_does_not_exist Test 64: error if parameter exists in the
   )
 })
 
-## Test 65: no error if the parameter exists in the dataset ----
-test_that("assert_param_does_not_exist Test 65: no error if the parameter exists in the dataset", {
+## Test 68: no error if the parameter exists in the dataset ----
+test_that("assert_param_does_not_exist Test 68: no error if the parameter exists in the dataset", {
   advs <- tibble::tribble(
     ~USUBJID, ~VSTESTCD, ~VSTRESN, ~VSSTRESU, ~PARAMCD, ~AVAL,
     "P01",    "WEIGHT",      80.1, "kg",      "WEIGHT",  80.1,
@@ -929,8 +957,8 @@ test_that("assert_param_does_not_exist Test 65: no error if the parameter exists
 })
 
 # assert_varval_list ----
-## Test 66: error if `arg` is not a list of var-value expressions ----
-test_that("assert_varval_list Test 66: error if `arg` is not a list of var-value expressions", {
+## Test 69: error if `arg` is not a list of var-value expressions ----
+test_that("assert_varval_list Test 69: error if `arg` is not a list of var-value expressions", {
   example_fun <- function(arg) {
     assert_varval_list(arg, accept_var = FALSE)
   }
@@ -940,8 +968,8 @@ test_that("assert_varval_list Test 66: error if `arg` is not a list of var-value
   )
 })
 
-## Test 67: error if `arg` is not a list of var-value expressions ----
-test_that("assert_varval_list Test 67: error if `arg` is not a list of var-value expressions", {
+## Test 70: error if `arg` is not a list of var-value expressions ----
+test_that("assert_varval_list Test 70: error if `arg` is not a list of var-value expressions", {
   example_fun <- function(arg) {
     assert_varval_list(arg, accept_var = TRUE)
   }
@@ -951,8 +979,8 @@ test_that("assert_varval_list Test 67: error if `arg` is not a list of var-value
   )
 })
 
-## Test 68: error if `required_elements` are missing from `arg` ----
-test_that("assert_varval_list Test 68: error if `required_elements` are missing from `arg`", {
+## Test 71: error if `required_elements` are missing from `arg` ----
+test_that("assert_varval_list Test 71: error if `required_elements` are missing from `arg`", {
   example_fun <- function(arg) {
     assert_varval_list(arg, required_elements = "DTHDOM")
   }
@@ -962,8 +990,8 @@ test_that("assert_varval_list Test 68: error if `required_elements` are missing 
   )
 })
 
-## Test 69: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_varval_list Test 69: no error if `arg` is NULL and optional is TRUE", {
+## Test 72: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_varval_list Test 72: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_varval_list(arg, optional = TRUE)
   }
@@ -973,8 +1001,8 @@ test_that("assert_varval_list Test 69: no error if `arg` is NULL and optional is
   )
 })
 
-## Test 70: error if `accept_expr` is TRUE and value is invalid ----
-test_that("assert_varval_list Test 70: error if `accept_expr` is TRUE and value is invalid", {
+## Test 73: error if `accept_expr` is TRUE and value is invalid ----
+test_that("assert_varval_list Test 73: error if `accept_expr` is TRUE and value is invalid", {
   example_fun <- function(arg) {
     assert_varval_list(arg, accept_expr = TRUE)
   }
@@ -984,8 +1012,8 @@ test_that("assert_varval_list Test 70: error if `accept_expr` is TRUE and value 
   )
 })
 
-## Test 71: error if `accept_expr` is FALSE and value is invalid ----
-test_that("assert_varval_list Test 71: error if `accept_expr` is FALSE and value is invalid", {
+## Test 74: error if `accept_expr` is FALSE and value is invalid ----
+test_that("assert_varval_list Test 74: error if `accept_expr` is FALSE and value is invalid", {
   example_fun <- function(arg) {
     assert_varval_list(arg, accept_expr = FALSE)
   }
@@ -995,8 +1023,8 @@ test_that("assert_varval_list Test 71: error if `accept_expr` is FALSE and value
   )
 })
 
-## Test 72: no error if an argument is a variable-value list ----
-test_that("assert_varval_list Test 72: no error if an argument is a variable-value list", {
+## Test 75: no error if an argument is a variable-value list ----
+test_that("assert_varval_list Test 75: no error if an argument is a variable-value list", {
   example_fun <- function(arg) {
     assert_varval_list(arg)
   }
@@ -1006,9 +1034,66 @@ test_that("assert_varval_list Test 72: no error if an argument is a variable-val
   )
 })
 
+# assert_expr_list ----
+## Test 76: error if `arg` is not a list of expressions ----
+test_that("assert_expr_list Test 76: error if `arg` is not a list of expressions", {
+  expect_error(
+    assert_expr_list(arg <- c("USUBJID", "PARAMCD", "VISIT")),
+    regexp = "`arg` must be a named list of expressions but it is a",
+    fixed = TRUE
+  )
+})
+
+## Test 77: error if `arg` is not a named list of expressions ----
+test_that("assert_expr_list Test 77: error if `arg` is not a named list of expressions", {
+  expect_error(
+    assert_expr_list(arg <- exprs(USUBJID, PARAMCD, NULL), named = TRUE),
+    regexp = "All elements of `arg` must be named.",
+    fixed = TRUE
+  )
+})
+
+## Test 78: error if `required_elements` are missing from `arg` ----
+test_that("assert_expr_list Test 78: error if `required_elements` are missing from `arg`", {
+  expect_error(
+    assert_expr_list(
+      arg <- exprs(DTHSEQ = AESEQ),
+      required_elements = "DTHDOM"),
+    regexp = "The following required elements are missing in `arg`: 'DTHDOM'",
+    fixed = TRUE
+  )
+})
+
+## Test 79: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_expr_list Test 79: no error if `arg` is NULL and optional is TRUE", {
+  expect_invisible(
+    assert_expr_list(NULL, optional = TRUE)
+  )
+})
+
+## Test 80: error if element is invalid ----
+test_that("assert_expr_list Test 80: error if element is invalid", {
+  expect_error(
+    assert_expr_list(arg <- exprs(DTHSEQ = !!mean)),
+    regexp = "The elements of the list arg must be expressions.",
+    fixed = TRUE
+  )
+})
+
+## Test 81: no error if argument is valid ----
+test_that("assert_expr_list Test 81: no error if argument is valid", {
+  expect_invisible(
+    assert_expr_list(arg <- exprs(
+      DTHDOM = "AE",
+      DTHSEQ = AESEQ,
+      DTHVAR = if_else(!is.na(AEDECOD), "AEDECOD", NA))
+      )
+  )
+})
+
 # assert_list_element ----
-## Test 73: no error if the elements fulfill a certain condition ----
-test_that("assert_list_element Test 73: no error if the elements fulfill a certain condition", {
+## Test 82: no error if the elements fulfill a certain condition ----
+test_that("assert_list_element Test 82: no error if the elements fulfill a certain condition", {
   expect_invisible(
     assert_list_element(
       list(
@@ -1022,8 +1107,8 @@ test_that("assert_list_element Test 73: no error if the elements fulfill a certa
   )
 })
 
-## Test 74: error if the elements do not fulfill the condition ----
-test_that("assert_list_element Test 74: error if the elements do not fulfill the condition", {
+## Test 83: error if the elements do not fulfill the condition ----
+test_that("assert_list_element Test 83: error if the elements do not fulfill the condition", {
   expect_error(
     assert_list_element(
       input <- list(
@@ -1040,23 +1125,23 @@ test_that("assert_list_element Test 74: error if the elements do not fulfill the
 })
 
 # assert_one_to_one ----
-## Test 75: error if there is a one to many mapping ----
-test_that("assert_one_to_one Test 75: error if there is a one to many mapping", {
+## Test 84: error if there is a one to many mapping ----
+test_that("assert_one_to_one Test 84: error if there is a one to many mapping", {
   expect_error(
     assert_one_to_one(admiral.test::admiral_dm, exprs(DOMAIN), exprs(USUBJID))
   )
 })
 
-## Test 76: error if there is a many to one mapping ----
-test_that("assert_one_to_one Test 76: error if there is a many to one mapping", {
+## Test 85: error if there is a many to one mapping ----
+test_that("assert_one_to_one Test 85: error if there is a many to one mapping", {
   expect_error(
     assert_one_to_one(admiral.test::admiral_dm, exprs(USUBJID), exprs(DOMAIN))
   )
 })
 
 # assert_date_var ----
-## Test 77: error if variable is not a date or datetime variable ----
-test_that("assert_date_var Test 77: error if variable is not a date or datetime variable", {
+## Test 86: error if variable is not a date or datetime variable ----
+test_that("assert_date_var Test 86: error if variable is not a date or datetime variable", {
   example_fun <- function(dataset, var) {
     var <- assert_symbol(enexpr(var))
     assert_date_var(dataset = dataset, var = !!var)
@@ -1077,18 +1162,18 @@ test_that("assert_date_var Test 77: error if variable is not a date or datetime 
 })
 
 # assert_date_vector ----
-## Test 78: returns error if input vector is not a date formatted ----
-test_that("assert_date_vector Test 78: returns error if input vector is not a date formatted", {
+## Test 87: returns error if input vector is not a date formatted ----
+test_that("assert_date_vector Test 87: returns error if input vector is not a date formatted", {
   expect_error(assert_date_vector("2018-08-23"))
 })
 
-## Test 79: returns invisible if input is date formatted ----
-test_that("assert_date_vector Test 79: returns invisible if input is date formatted", {
+## Test 88: returns invisible if input is date formatted ----
+test_that("assert_date_vector Test 88: returns invisible if input is date formatted", {
   expect_invisible(assert_date_vector(as.Date("2022-10-25")))
 })
 
-## Test 80: no error if `arg` is NULL and optional is TRUE ----
-test_that("assert_date_vector Test 80: no error if `arg` is NULL and optional is TRUE", {
+## Test 89: no error if `arg` is NULL and optional is TRUE ----
+test_that("assert_date_vector Test 89: no error if `arg` is NULL and optional is TRUE", {
   example_fun <- function(arg) {
     assert_date_vector(arg, optional = TRUE)
   }
@@ -1098,8 +1183,8 @@ test_that("assert_date_vector Test 80: no error if `arg` is NULL and optional is
   )
 })
 
-## Test 81: error if `arg` is NULL and optional is FALSE ----
-test_that("assert_date_vector Test 81: error if `arg` is NULL and optional is FALSE", {
+## Test 90: error if `arg` is NULL and optional is FALSE ----
+test_that("assert_date_vector Test 90: error if `arg` is NULL and optional is FALSE", {
   example_fun <- function(arg) {
     assert_date_vector(arg, optional = FALSE)
   }
@@ -1113,22 +1198,22 @@ test_that("assert_date_vector Test 81: error if `arg` is NULL and optional is FA
 
 
 # assert_atomic_vector ----
-## Test 82: error if input is not atomic vector ----
-test_that("assert_atomic_vector Test 82: error if input is not atomic vector", {
+## Test 91: error if input is not atomic vector ----
+test_that("assert_atomic_vector Test 91: error if input is not atomic vector", {
   x <- list("a", "a", "b", "c", "d", "d", 1, 1, 4)
   expect_error(assert_atomic_vector(x))
 })
 
 # assert_same_type ----
-## Test 83: no error if same type ----
-test_that("assert_same_type Test 83: no error if same type", {
+## Test 92: no error if same type ----
+test_that("assert_same_type Test 92: no error if same type", {
   true_value <- "Y"
   false_value <- "N"
   expect_invisible(assert_same_type(true_value, false_value))
 })
 
-## Test 84: error if different type ----
-test_that("assert_same_type Test 84: error if different type", {
+## Test 93: error if different type ----
+test_that("assert_same_type Test 93: error if different type", {
   true_value <- "Y"
   false_value <- "N"
   missing_value <- 0
@@ -1147,8 +1232,8 @@ test_that("assert_same_type Test 84: error if different type", {
   )
 })
 
-## Test 85: works as intended ----
-test_that("assert_same_type Test 85: works as intended", {
+## Test 94: works as intended ----
+test_that("assert_same_type Test 94: works as intended", {
   expect_equal(
     valid_time_units(),
     c("years", "months", "days", "hours", "minutes", "seconds")